+++ scripts/generate-ts.sh
@@ -7,6 +7,10 @@ OPENAPI_GENERATED_TS_PATH=$APP_PATH/src/types/generated
 SRC_FILES=(
     "legacy/ojp-v1/ojp-lir-request.yml"
     "legacy/ojp-v1/ojp-lir-response.yml"
+    
+    "legacy/ojp-v1/ojp-ser-request.yml"
+    "legacy/ojp-v1/ojp-ser-response.yml"
+    
     "legacy/ojp-v1/ojp-shared.yml"
     "legacy/ojp-v1/ojp-tir-response.yml"
     "legacy/ojp-v1/ojp-trip.yml"