+++ src/types/openapi-dependencies.ts
@@ -99,8 +99,14 @@ const MapArrayTags: Record<string, boolean> = {
   'places.place': true,
   'linkProjection.position': true,
 
+  // OJP v1
+  'places.location': true,
+
+  // PSituation Refs
   'situationFullRefs.situationFullRef': true,
   'situations.ptSituation': true,
+
+  // PtSituation OJP v2
   'ptSituation.validityPeriod': true,
   'publishingActions.publishingAction': true,
   'publishingAction.passengerInformationAction': true,
@@ -110,6 +116,18 @@ const MapArrayTags: Record<string, boolean> = {
   'textualContent.recommendationContent': true,
   'textualContent.remarkContent': true,
   'textualContent.infoLink': true,
+  // technically these are also arrays - but the OJP repsonse never returned it as such
+  // 'descriptionContent.descriptionText': true,
+  // 'consequenceContent.consequenceText': true,
+  // 'recommendationContent.recommendationText': true,
+  // 'remarkContent.remark': true,
+  
+  // PtSituation OJP v1
+  'callAtStop.situationFullRef': true,
+  'ptSituation.description': true,
+  'ptSituation.detail': true,
+  'stopPoints.affectedStopPoint': true,
+  'vehicleJourneys.affectedVehicleJourney': true,
 
   'itModeAndModeOfOperation.personalModeOfOperation': true,
 
@@ -133,10 +151,13 @@ const MapArrayTags: Record<string, boolean> = {
   'legAlight.expectedDepartureOccupancy': true,
 
   // LIR Response
-  'place.mode': true,
   'OJPLocationInformationDelivery.placeResult': true,
+  'place.mode': true,
+  'place.attribute': true,
   // OJP 1.0
   'OJPLocationInformationDelivery.location': true,
+  'location.mode': true,
+  'location.attribute': true,
   
   // LIR Response POI
   'pointOfInterest.pointOfInterestCategory': true,
@@ -144,14 +165,15 @@ const MapArrayTags: Record<string, boolean> = {
   'pointOfInterestCategory.pointOfInterestClassification': true,
   'pOIAdditionalInformation.pOIAdditionalInformation': true,
 
-  // SER Response
-  'stopEvent.previousCall': true,
-  'stopEvent.onwardCall': true,
-
   // Fare Response
   'OJPFareDelivery.fareResult': true,
   'fareResult.tripFareResult': true, // TODO - this is not in the current schema (v2)
   'tripFareResult.fareProduct': true,
+
+  // StopEvent Response
+  'OJPStopEventDelivery.stopEventResult': true,
+  'stopEvent.previousCall': true,
+  'stopEvent.onwardCall': true,
 
   // TripInfoResult
   'OJPTripInfoDelivery.tripInfoResult': true,
@@ -162,6 +184,8 @@ const MapArrayTags: Record<string, boolean> = {
   // TripRefineResult
   'OJPTripRefineDelivery.tripResult': true,
 };
+
+// This will be deleted in the new update
 
 // TODO - this should be generated
 // Declare object properties that are / should be serialised to XML